Remove filter module and move VideoFilter
