+++ buildSrc/src/main/java/DependencyHandler.kt
@@ -108,7 +108,6 @@ fun DependencyHandler.moduleDependencies() {
     FEATURE_MY_PROFILE
     FEATURE_SETTING
     FEATURE_CAMERA_MEDIA
-    FEATURE_FILTER
 }
 
 
@@ -157,5 +156,3 @@ val DependencyHandler.FEATURE_SETTING
 val DependencyHandler.FEATURE_CAMERA_MEDIA
     get() = implementation(project(mapOf("path" to ":feature:cameramedia")))
 
-val DependencyHandler.FEATURE_FILTER
-    get() = implementation(project(mapOf("path" to ":feature:filter")))
