--- feature/filter/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("plugin.android-common")
-}
-
-
-dependencies {
-    COMMON_THEME
-    COMMON_COMPOSABLE
-    DOMAIN
-    DATA
-    CORE
-    implementation(project(":feature:mp4compose"))
-}
