+++ settings.gradle.kts
@@ -29,7 +29,6 @@ include(":feature:friends")
 include(":feature:myprofile")
 include(":feature:setting")
 include(":feature:cameramedia")
-include(":feature:filter")
 include(":feature:cameracapture")
 include(":feature:video-trimmer")
 include(":feature:mp4compose")
